+++ NEWS.md
@@ -15,7 +15,6 @@ a series of dates. Changed interface of `enw_preprocess_data()` to pass `...` to
 - Refined the use of data ordering throughout the preprocessing functions. See #147 by @seabbs.
 - Skipped tests that use `cmdstan` locally to improve the developer/contributor experience. See #147 by @seabbs and @adrian-lison.
 - Added a basic simulator function for missing reference data. See #147 by @seabbs and @adrian-lison.
-- Added a basic simulator function for missing reference data. See #147 by @seabbs and @adrian-lison.
 - Added support for right hand side interactions as syntax sugar for random effects. This allows the specification of, for example, independent random effects by day for each strata of another variable. See #169 by @seabbs.
 
 ## Model
@@ -27,6 +26,10 @@ a series of dates. Changed interface of `enw_preprocess_data()` to pass `...` to
 - Added support for supplying missingness model parameters to the model as well as optional priors and effect estimation. See #138 by @seabbs and @adrian-lison.
 - Refactored model generated quantities to be functional. See #138 by @seabbs and @adrian-lison.
 - Added support for modelling missing reference dates to the likelihood. See #147 by @seabbs and @adrian-lison.
+- Added additional functionality to `delay_group_lmpf` to support modelling observations missing reference dates. Also updated the generated quantities to support this mode. See #147 by @seabbs and @adrian-lison based on #64 by @adrian-lison.
+- Added a flexible expectation process on the growth rate scale. The default expectation model has been updated to a group-wise random walk on the growth rate. See #152 by @seabbs and @adrian-lison.
+- Added a deterministic  renewal equation, and latent reporting process. See #152 by @seabbs and @adrian-lison.
+- Added support for no intercept in the expectation model and more general formula support to enable this as a feature in other modules going forward. See #170 by @seabbs.
 
 ## Documentation
 - Removed explicit links to authors and issues in the `NEWS.md` file. See #132 by @choi-hannah.